+++ examples/meshlinkapp.c
@@ -3,7 +3,7 @@
 #include "meshlink/meshlink.h"
 
 void handle_recv_data(meshlink_handle_t *mesh, meshlink_node_t *source, void *data, size_t len) {
-	printf("Received %zu bytes from %s: %s\n", len, source->name, (char*)data);
+	printf("Received %lu bytes from %s: %s\n", (unsigned long) len, source->name, (char*)data);
 }
 
 int main(int argc , char **argv){
